+++ crates/forgeql-core/src/filter.rs
@@ -790,30 +790,35 @@ mod tests {
                 text: "void setup() {".into(),
                 marker: None,
                 node_id: None,
+                node_offset: None,
             },
             SourceLine {
                 line: 11,
                 text: "    // TODO: fix this".into(),
                 marker: None,
                 node_id: None,
+                node_offset: None,
             },
             SourceLine {
                 line: 12,
                 text: "    int x = 42;".into(),
                 marker: None,
                 node_id: None,
+                node_offset: None,
             },
             SourceLine {
                 line: 13,
                 text: "    // FIXME: needs review".into(),
                 marker: None,
                 node_id: None,
+                node_offset: None,
             },
             SourceLine {
                 line: 14,
                 text: "}".into(),
                 marker: None,
                 node_id: None,
+                node_offset: None,
             },
         ]
     }